LayoutTests/inspector/css/overrideUserPreferenceMatchedStyles-expected.txt

Tests for matched styles when forcing media features with CSSManager.overrideUserPreference().


== Running test suite: CSSManager.overrideUserPreference
-- Running test case: Override PrefersReducedMotion: On
Override preference PrefersReducedMotion to On
PASS: --test-prefers-reduced-motion value matches on
Clear override for PrefersReducedMotion

-- Running test case: Override PrefersContrast: On
Override preference PrefersContrast to On
PASS: --test-prefers-contrast value matches on
Clear override for PrefersContrast

-- Running test case: PrefersColorScheme.Dark
Override preference PrefersColorScheme to Dark
PASS: --test-prefers-color-scheme value matches dark
Clear override for PrefersColorScheme
